Moved DestoryTask from Powertask to BasicTask
